pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pytelegrambotapi>=4.26.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
